Break the pricing models into separate files
